create Section component containing exportable Section elements
